fix format

+++ JxbScaleButton/JxbScaleButton/JxbScaleButton.h
@@ -12,7 +12,7 @@
 @property(nonatomic,strong)NSString     *strCommon;//按钮可用时的文本
 @property(nonatomic,strong)NSString     *strPrefix;//倒计时前缀
 @property(nonatomic,strong)NSString     *strSuffix;//倒计时后缀
-@property(nonatomic,assign)NSInteger    indexStart;//开始从几倒计时
+@property(nonatomic,assign)int          indexStart;//开始从几倒计时
 @property(nonatomic,strong)UIColor      *colorDisable;//倒计时的背景颜色
 @property(nonatomic,strong)UIColor      *colorCommon;//按钮可用时的背景颜色
 @property(nonatomic,strong)UIColor      *colorTitle;//文本颜色
